refactoring: Namespaces fixed for rules
